별점 소수점 수정

+++ src/main/java/com/kkosoonnae/jpa/projection/StoreDetailViewProjection.java
@@ -1,9 +1,7 @@
 package com.kkosoonnae.jpa.projection;
 
-import com.kkosoonnae.store.dto.StoreDetailWithImageResponseDto;
 import lombok.Builder;
 
-import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -18,7 +16,7 @@ public record StoreDetailViewProjection(
         LocalTime openingTime,
         LocalTime closingTime,
         List<String> img,
-        Double averageScope,
+        Integer averageScope,
         Long totalLikeStore
 ) {
     public StoreDetailViewProjection(Integer storeNo, String storeName, String content, String phone,
